# optimize replace! on arrays