@KristofferC 2020-02-26
LLVM does some runtime aliasing checks already but maybe it doesn't realize it is safe to run the vectorized version if the pointers are identical? Or maybe we have some higher level knowledge that LLVM cannot know from what is passed to it?