@rfourquet 2020-04-13
It's a trivial PR, so I will merge in a few days if no objection.